+++ backend/src/main/java/com/dubu/backend/plan/dto/response/RouteSearchResponseDto.java
@@ -1,4 +1,4 @@
-package com.dubu.backend.plan.dto;
+package com.dubu.backend.plan.dto.response;
 
 import java.util.List;
